CORE-2066 Move `ContactSupport` and `GridLabelValue` into ./common/
